rename files and directory to conform to ruby gem naming conventions
