+++ lib/xqueue_ruby/xqueue_submission.rb
